+++ Tests/StreamChatTests/Query/MessageSearchQuery_Tests.swift
@@ -6,16 +6,29 @@
 import XCTest
 
 final class MessageSearchFilterScope_Tests: StressTestCase {
-    func test_withAttachments() {
-        // Declare attachment types
+    func test_withAttachments_givenAttachmentTypes() {
         let attachmentTypes: Set<AttachmentType> = [.image, .video]
-        
-        // Build a filter
+
         let filter: Filter<MessageSearchFilterScope> = .withAttachments(attachmentTypes)
-        
-        // Assert correct filter is built
-        XCTAssertEqual(filter.key, FilterKey<MessageSearchFilterScope, AttachmentType>.hasAttachmentsOfType.rawValue)
-        XCTAssertEqual(filter.operator, FilterOperator.in.rawValue)
+
+        XCTAssertEqual(filter.key, "attachments.type")
+        XCTAssertEqual(filter.operator, "$in")
         XCTAssertEqual(Set(filter.value as! [AttachmentType]), attachmentTypes)
+    }
+
+    func test_withAttachments() {
+        let filter: Filter<MessageSearchFilterScope> = .withAttachments
+
+        XCTAssertEqual(filter.key, "attachments")
+        XCTAssertEqual(filter.operator, "$exists")
+        XCTAssertEqual(filter.value as! Bool, true)
+    }
+
+    func test_withoutAttachments() {
+        let filter: Filter<MessageSearchFilterScope> = .withoutAttachments
+
+        XCTAssertEqual(filter.key, "attachments")
+        XCTAssertEqual(filter.operator, "$exists")
+        XCTAssertEqual(filter.value as! Bool, false)
     }
 }